move padlfile commands out of project
